# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Nexodus API Server.",
        "title": "Nexodus API",
        "contact": {
            "name": "The Nexodus Authors",
            "url": "https://github.com/nexodus-io/nexodus/issues"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/ca/sign": {
            "post": {
                "description": "Signs a certificate signing request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CA"
                ],
                "summary": "Signs a certificate signing request",
                "operationId": "SignCSR",
                "parameters": [
                    {
                        "description": "Certificate signing request",
                        "name": "CertificateSigningRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CertificateSigningRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CertificateSigningResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CertificateSigningRequest": {
            "type": "object",
            "properties": {
                "duration": {
                    "description": "Requested 'duration' (i.e. lifetime) of the Certificate. Note that the\nissuer may choose to ignore the requested duration, just like any other\nrequested attribute.\n+optional",
                    "type": "string"
                },
                "is_ca": {
                    "description": "Requested basic constraints isCA value. Note that the issuer may choose\nto ignore the requested isCA value, just like any other requested attribute.\n\nNOTE: If the CSR in the `Request` field has a BasicConstraints extension,\nit must have the same isCA value as specified here.\n\nIf true, this will automatically add the `cert sign` usage to the list\nof requested `usages`.\n+optional",
                    "type": "boolean"
                },
                "request": {
                    "description": "The PEM-encoded X.509 certificate signing request to be submitted to the\nissuer for signing.\n\nIf the CSR has a BasicConstraints extension, its isCA attribute must\nmatch the `isCA` value of this CertificateRequest.\nIf the CSR has a KeyUsage extension, its key usages must match the\nkey usages in the `usages` field of this CertificateRequest.\nIf the CSR has a ExtKeyUsage extension, its extended key usages\nmust match the extended key usages in the `usages` field of this\nCertificateRequest.",
                    "type": "string",
                    "example": "-----BEGIN CERTIFICATE REQUEST-----(...)-----END CERTIFICATE REQUEST-----"
                },
                "usages": {
                    "description": "Requested key usages and extended key usages.\n\nNOTE: If the CSR in the `Request` field has uses the KeyUsage or\nExtKeyUsage extension, these extensions must have the same values\nas specified here without any additional values.\n\nIf unset, defaults to `digital signature` and `key encipherment`.\n+optional",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.KeyUsage"
                    }
                }
            }
        },
        "models.CertificateSigningResponse": {
            "type": "object",
            "properties": {
                "ca": {
                    "description": "The PEM encoded X.509 certificate of the signer, also known as the CA\n(Certificate Authority).\nThis is set on a best-effort basis by different issuers.\nIf not set, the CA is assumed to be unknown/not available.\n+optional",
                    "type": "string",
                    "example": "-----BEGIN CERTIFICATE-----(...)-----END CERTIFICATE-----"
                },
                "certificate": {
                    "description": "The PEM encoded X.509 certificate resulting from the certificate\nsigning request.\nIf not set, the CertificateRequest has either not been completed or has\nfailed. More information on failure can be found by checking the\n`conditions` field.\n+optional",
                    "type": "string",
                    "example": "-----BEGIN CERTIFICATE-----(...)-----END CERTIFICATE-----"
                }
            }
        },
...
        "models.InternalServerError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something bad"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
...
        "models.KeyUsage": {
            "type": "string",
            "enum": [
                "signing",
                "digital signature",
                "content commitment",
                "key encipherment",
                "key agreement",
                "data encipherment",
                "cert sign",
                "crl sign",
                "encipher only",
                "decipher only",
                "any",
                "server auth",
                "client auth",
                "code signing",
                "email protection",
                "s/mime",
                "ipsec end system",
                "ipsec tunnel",
                "ipsec user",
                "timestamping",
                "ocsp signing",
                "microsoft sgc",
                "netscape sgc"
            ],
            "x-enum-varnames": [
                "UsageSigning",
                "UsageDigitalSignature",
                "UsageContentCommitment",
                "UsageKeyEncipherment",
                "UsageKeyAgreement",
                "UsageDataEncipherment",
                "UsageCertSign",
                "UsageCRLSign",
                "UsageEncipherOnly",
                "UsageDecipherOnly",
                "UsageAny",
                "UsageServerAuth",
                "UsageClientAuth",
                "UsageCodeSigning",
                "UsageEmailProtection",
                "UsageSMIME",
                "UsageIPsecEndSystem",
                "UsageIPsecTunnel",
                "UsageIPsecUser",
                "UsageTimestamping",
                "UsageOCSPSigning",
                "UsageMicrosoftSGC",
                "UsageNetscapeSGC"
            ]
        },
...
        "models.ValidationError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something bad"
                },
                "field": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://auth.try.nexodus.127.0.0.1.nip.io/",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "user": " Grants read and write access to resources owned by this user"
            }
        }
    }
}